Change architecture #JC19-326
