+++ apps/vanilla/src/constants/localStorage.ts
@@ -1,4 +1,4 @@
 /** Keys for which information is located in the local storage. */
-export namespace LocalStorageKeys {
+export namespace LocalStorageKey {
   export const SORT_SETTINGS = 'sortSettings';
 }
