+++ apps/vanilla/src/services/api/anime.ts
