+++ apps/vanilla/src/services/domain/localStorage.ts
@@ -3,7 +3,7 @@
  * @param key The key by which the value is stored.
  * @returns If there are no value, then null otherwise the value object.
  */
-export function getLocalStorage<T>(key: string): T | null {
+export function getValueFromLocalStorage<T>(key: string): T | null {
   const localValue = localStorage.getItem(key);
   if (localValue !== null) {
     return JSON.parse(localValue);
@@ -16,7 +16,7 @@ export function getLocalStorage<T>(key: string): T | null {
  * @param key The key by which the value is stored.
  * @param value Some value that needs to be stored.
  */
-export function setLocalStorage<T>(key: string, value: T): void {
+export function setValueToLocalStorage<T>(key: string, value: T): void {
   if (value !== undefined) {
     localStorage.setItem(key, JSON.stringify(value));
   }
